Move drawer to separate component
