+++ src/components/header.js
@@ -1,105 +1,13 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Navbar, Nav } from 'react-bootstrap';
 import Logo from '../images/gatsby-icon.png';
 import { Link } from 'gatsby';
 import { NavDropdown } from 'react-bootstrap';
 import { externalLinks } from '../common/externalLinks';
-import Button from '@material-ui/core/Button';
-import MenuIcon from '@material-ui/icons/Menu';
-import { makeStyles } from '@material-ui/core/styles';
-import Drawer from '@material-ui/core/Drawer';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemText from '@material-ui/core/ListItemText';
-import List from '@material-ui/core/List';
+import Drawer from './drawer';
 import './header.css';
 
-const useStyles = makeStyles({
-  list: {
-    width: 250,
-  },
-  listItemLink: {
-    textDecoration: 'none',
-  },
-});
-
 const Header = () => {
-  const classes = useStyles();
-  const [isDrawerToggled, setIsDrawerToggled] = useState(false);
-
-  const toggleDrawer = (open) => (event) => {
-    if (
-      event.type === 'keydown' &&
-      (event.key === 'Tab' || event.key === 'Shift')
-    ) {
-      return;
-    }
-    setIsDrawerToggled(open);
-  };
-
-  const getExternalListItem = ({ link, key, text }) => (
-    <a href={externalLinks[link]}>
-      <ListItem key={key}>
-        <ListItemText primary={text} />
-      </ListItem>
-    </a>
-  );
-
-  const getCommunityListItems = () =>
-    [
-      { link: 'discord', key: 'discord', text: 'Discord' },
-      { link: 'reddit', key: 'reddit', text: 'Reddit' },
-      { link: 'twitter', key: 'twitter', text: 'Twitter' },
-    ].map((listItemInfo) => getExternalListItem(listItemInfo));
-
-  const getAboutListItems = () => {
-    const internalLinks = [
-      { link: '/documentation', key: 'documentation', text: 'Documentation' },
-      { link: '/', key: 'work-with-us', text: 'Work with Us' },
-      { link: '/', key: 'contact-us', text: 'Contact Us' },
-    ].map((aboutListItemInfo) => (
-      <Link to={aboutListItemInfo.link}>
-        <ListItem key={aboutListItemInfo.key}>
-          <ListItemText primary={aboutListItemInfo.text} />
-        </ListItem>
-      </Link>
-    ));
-    const externalLinks = [
-      getExternalListItem({ link: 'github', key: 'github', text: 'GitHub' }),
-    ];
-
-    return internalLinks.concat(externalLinks);
-  };
-
-  const drawerList = () => (
-    <div
-      className={classes.list}
-      role="presentation"
-      onClick={toggleDrawer(false)}
-      onKeyDown={toggleDrawer(false)}
-    >
-      <List>
-        <Link to="/funding">
-          <ListItem button key={'Funding'}>
-            <ListItemText primary={'Funding'} />
-          </ListItem>
-        </Link>
-      </List>
-      <List>
-        <ListItem key="Community">
-          <ListItemText primary="Community" />
-        </ListItem>
-        {getCommunityListItems()}
-      </List>
-      <List>
-        <ListItem key="About">
-          <ListItemText primary="About" />
-        </ListItem>
-        {/* {getCommunityListItems()} */}
-        {getAboutListItems()}
-      </List>
-    </div>
-  );
-
   return (
     <header>
       <Navbar fixed="top" bg="white" expand="lg">
@@ -151,16 +59,7 @@ const Header = () => {
             Use Basic Credit
           </Nav.Link>
         </Navbar>
-        <Button onClick={toggleDrawer(true)}>
-          <MenuIcon className="mobile-nav" />
-        </Button>
-        <Drawer
-          anchor="right"
-          open={isDrawerToggled}
-          onClose={toggleDrawer(false)}
-        >
-          {drawerList()}
-        </Drawer>
+        <Drawer />
       </Navbar>
     </header>
   );
